+++ wdm/app/src/main/java/com/versilistyson/welldone/data/repository/SensorRepositoryImpl.kt
@@ -24,9 +24,9 @@ class SensorRepositoryImpl @Inject constructor(
 
     val store =
         StoreBuilder
-            .fromNonFlow<StoreKey.SensorsKey, List<SensorData>> {storekey ->
+            .fromNonFlow<StoreKey.SensorsKey, List<SensorData>> { storekey ->
                 val sensors = mutableListOf<SensorData>()
-                val sensorSet = mutableSetOf<Int>()
+                val sensorSet = mutableSetOf<Long>()
                 remoteDataSource.getSensors(storekey).body()?.forEach {
                     if(!sensorSet.contains(it.sensorId)) { //issue in backend where it duplicates assigned sensors
                         sensors.add(it.map())
